+++ R/list_animal_ids.R
@@ -1,5 +1,9 @@
 #' List all available animal ids
 #'
+#' @param api Logical. If `TRUE`, the function will use the API to access ETN.
+#' @param connection `r lifecycle::badge("deprecated")` A connection to the ETN
+#'   database. This argument is no longer used. You will be prompted for
+#'   credentials instead.
 #'
 #' @return A vector of all unique `id_pk` present in `common.animal_release`.
 #'