+++ tests/testthat/test-write_dwc.R
@@ -1,11 +1,9 @@
-con <- connect_to_etn()
-
-test_that("write_dwc() can write csv files to a path", {
+test_that("[SQL] write_dwc() can write csv files to a path", {
   out_dir <- file.path(tempdir(), "dwc")
   unlink(out_dir, recursive = TRUE)
   dir.create(out_dir)
   suppressMessages(
-    write_dwc(con, animal_project_code = "2014_demer", directory = out_dir)
+    write_dwc(animal_project_code = "2014_demer", directory = out_dir, api = FALSE)
   )
 
   expect_identical(
@@ -14,18 +12,18 @@ test_that("write_dwc() can write csv files to a path", {
   )
 })
 
-test_that("write_dwc() can return data as list of tibbles rather than files", {
+test_that("[SQL] write_dwc() can return data as list of tibbles rather than files", {
   result <- suppressMessages(
-    write_dwc(con, animal_project_code = "2014_demer", directory = NULL)
+    write_dwc(animal_project_code = "2014_demer", directory = NULL, api = FALSE)
   )
 
   expect_identical(names(result), "dwc_occurrence")
   expect_s3_class(result$dwc_occurrence, "tbl")
 })
 
-test_that("write_dwc() returns the expected Darwin Core terms as columns", {
+test_that("[SQL] write_dwc() returns the expected Darwin Core terms as columns", {
   result <- suppressMessages(
-    write_dwc(con, animal_project_code = "2014_demer", directory = NULL)
+    write_dwc(animal_project_code = "2014_demer", directory = NULL, api = FALSE)
   )
 
   expect_identical(
@@ -62,4 +60,68 @@ test_that("write_dwc() returns the expected Darwin Core terms as columns", {
       "kingdom"
     )
   )
+})
+
+# cached version of API response for 2014_demer
+vcr::use_cassette("2014_demer_dwc", {
+  # Request without writing to disk
+  demer_dwc <- write_dwc(animal_project_code = "2014_demer",
+                         directory = NULL,
+                         api = TRUE)
+
+  test_that("[API] write_dwc() can write csv files to a path", {
+    out_dir <- file.path(tempdir(), "dwc")
+    unlink(out_dir, recursive = TRUE)
+    dir.create(out_dir)
+    # Request that writes to disk
+    suppressMessages(write_dwc(
+      animal_project_code = "2014_demer",
+      directory = out_dir,
+      api = TRUE
+    ))
+
+    expect_identical(list.files(out_dir, pattern = "*.csv"), "dwc_occurrence.csv")
+  })
+
+  test_that("[API] write_dwc() can return data as list of tibbles rather than files", {
+    expect_identical(names(demer_dwc), "dwc_occurrence")
+    expect_s3_class(demer_dwc$dwc_occurrence, "tbl")
+  })
+
+  test_that("[API] write_dwc() returns the expected Darwin Core terms as columns", {
+    expect_identical(
+      colnames(demer_dwc$dwc_occurrence),
+      c(
+        "type",
+        "license",
+        "rightsHolder",
+        "datasetID",
+        "institutionCode",
+        "collectionCode",
+        "datasetName",
+        "basisOfRecord",
+        "dataGeneralizations",
+        "occurrenceID",
+        "sex",
+        "lifeStage",
+        "occurrenceStatus",
+        "organismID",
+        "organismName",
+        "eventID",
+        "parentEventID",
+        "eventDate",
+        "samplingProtocol",
+        "eventRemarks",
+        "locationID",
+        "locality",
+        "decimalLatitude",
+        "decimalLongitude",
+        "geodeticDatum",
+        "coordinateUncertaintyInMeters",
+        "scientificNameID",
+        "scientificName",
+        "kingdom"
+      )
+    )
+  })
 })