+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Deprecated functions `get_deployments()`, `get_detections()`, `get_projects()`, `get_receivers()`, `list_network_project_codes()` and `list_tag_ids()` are now end of life and no longer included in etn 3.0.0
+* `write_dwc()` now invisibly returns a list of data.frames even when writing out to a file (with `directory` set to a path)
+